requirements.txt
black==22.10.0
inquirer
pyfiglet
pre-commit
a2wsgi==1.6.0
keyboard==0.13.5
flask==2.2.2
flask-cors==3.0.10
python-dotenv==0.21.0
python-telegram-bot==20.0
